backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-allauth>=0.57,<1.0
dj-rest-auth>=5.0,<6.0
python-decouple>=3.8,<4.0
requests>=2.31,<3.0
django-filter>=23.0,<24.0
openpyxl>=3.1,<4.0
psycopg2-binary>=2.9,<3.0
